Rename _pytest._nodeid to _pytest.nodeid
The module is already under the _pytest private package, so an extra
leading underscore on the module name itself is redundant.
