+++ doc/en/conf.py
@@ -90,10 +90,10 @@ nitpick_ignore = [
     ("py:class", "_pytest.recwarn.WarningsChecker"),
     ("py:class", "_pytest.reports.BaseReport"),
     ("py:class", "_pytest.reports._WithNodeId"),
-    ("py:class", "_pytest._nodeid.NodeId"),
-    ("py:class", "_pytest._nodeid.CollectionNodeId"),
-    ("py:class", "_pytest._nodeid.ItemNodeId"),
-    ("py:class", "_pytest._nodeid.OpaqueNodeId"),
+    ("py:class", "_pytest.nodeid.NodeId"),
+    ("py:class", "_pytest.nodeid.CollectionNodeId"),
+    ("py:class", "_pytest.nodeid.ItemNodeId"),
+    ("py:class", "_pytest.nodeid.OpaqueNodeId"),
     # Sphinx bugs(?)
     ("py:class", "RewriteHook"),
     # Undocumented third parties
